art: Rename base and CRUD mixins to include underscore prefix for clarity
