+++ src/api/core/models/__init__.py
@@ -1,3 +1,3 @@
 # -*- coding: utf-8 -*-
 
-from .base import *
+from ._base import *
