+++ src/api/core/models/mixins/__init__.py
@@ -1,4 +1,4 @@
 # -*- coding: utf-8 -*-
 
-from .base import *
-from .crud import *
+from ._base import *
+from ._crud import *
