+++ src/api/core/models/mixins/async_/_delete.py
@@ -18,7 +18,7 @@ if config.db.dialect == "postgresql":
 from api.core.exceptions import EmptyValueError, ForeignKeyError
 from api.logger import logger
 
-from .read import AsyncReadMixin
+from ._read import AsyncReadMixin
 
 
 @declarative_mixin
