+++ src/api/core/models/mixins/sync/_delete.py
@@ -15,7 +15,7 @@ if config.db.dialect == "postgresql":
 from api.core.exceptions import EmptyValueError, ForeignKeyError
 from api.logger import logger
 
-from .read import ReadMixin
+from ._read import ReadMixin
 
 
 @declarative_mixin
