+++ src/api/core/models/mixins/sync/_read.py
