+++ src/api/databases/rdb/_check.py
@@ -11,7 +11,7 @@ from api.core.constants import WarnEnum
 from api.config import config
 from api.logger import logger
 
-from .create import async_create_db, create_db
+from ._create import async_create_db, create_db
 
 
 ## Async
