+++ src/api/databases/rdb/_close.py
